+++ examples/hot_reload_tide/src/main.rs
@@ -8,6 +8,7 @@ use notify::{
     event::ModifyKind,
     Error, Event, EventKind, RecommendedWatcher, RecursiveMode, Watcher,
 };
+use std::path::Path;
 use std::sync::{Arc, Mutex};
 use tide::{Body, Response};
 
@@ -28,11 +29,11 @@ async fn main() -> tide::Result<()> {
 
     // We listen to file changes by giving Notify
     // a function that will get called when events happen
-    let mut watcher: RecommendedWatcher =
+    let mut watcher =
         // To make sure that the config lives as long as the function
         // we need to move the ownership of the config inside the function
         // To learn more about move please read [Using move Closures with Threads](https://doc.rust-lang.org/book/ch16-01-threads.html?highlight=move#using-move-closures-with-threads)
-        Watcher::new_immediate(move |result: Result<Event, Error>| {
+        RecommendedWatcher::new(move |result: Result<Event, Error>| {
             let event = result.unwrap();
 
             if event.kind == EventKind::Modify(ModifyKind::Any) {
@@ -43,7 +44,7 @@ async fn main() -> tide::Result<()> {
             }
         })?;
 
-    watcher.watch(CONFIG_PATH, RecursiveMode::Recursive)?;
+    watcher.watch(Path::new(CONFIG_PATH), RecursiveMode::Recursive)?;
 
     // We set up a web server using [Tide](https://github.com/http-rs/tide)
     let mut app = tide::with_state(config);