+++ src/lib.rs
@@ -4,7 +4,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! notify = "5.0.0-pre.9"
+//! notify = "5.0.0-pre.10"
 //! ```
 //!
 //! ## Serde
@@ -12,7 +12,7 @@
 //! Events are serialisable via [serde] if the `serde` feature is enabled:
 //!
 //! ```toml
-//! notify = { version = "5.0.0-pre.9", features = ["serde"] }
+//! notify = { version = "5.0.0-pre.10", features = ["serde"] }
 //! ```
 //!
 //! [serde]: https://serde.rs
@@ -20,11 +20,12 @@
 //! # Examples
 //!
 //! ```
+//! # use std::path::Path;
 //! use notify::{Watcher, RecommendedWatcher, RecursiveMode, Result};
 //!
 //! fn main() -> Result<()> {
 //!     // Automatically select the best implementation for your platform.
-//!     let mut watcher: RecommendedWatcher = Watcher::new_immediate(|res| {
+//!     let mut watcher = notify::recommended_watcher(|res| {
 //!         match res {
 //!            Ok(event) => println!("event: {:?}", event),
 //!            Err(e) => println!("watch error: {:?}", e),
@@ -33,7 +34,7 @@
 //!
 //!     // Add a path to be watched. All files and directories at that path and
 //!     // below will be monitored for changes.
-//!     watcher.watch(".", RecursiveMode::Recursive)?;
+//!     watcher.watch(Path::new("."), RecursiveMode::Recursive)?;
 //!
 //!     Ok(())
 //! }
@@ -48,10 +49,11 @@
 //!
 //! ```
 //! # use notify::{Watcher, RecommendedWatcher, RecursiveMode, Result};
+//! # use std::path::Path;
 //! # use std::time::Duration;
 //! # fn main() -> Result<()> {
 //! # // Automatically select the best implementation for your platform.
-//! # let mut watcher: RecommendedWatcher = Watcher::new_immediate(|res| {
+//! # let mut watcher = RecommendedWatcher::new(|res| {
 //! #     match res {
 //! #        Ok(event) => println!("event: {:?}", event),
 //! #        Err(e) => println!("watch error: {:?}", e),
@@ -60,7 +62,7 @@
 //!
 //! # // Add a path to be watched. All files and directories at that path and
 //! # // below will be monitored for changes.
-//! # watcher.watch(".", RecursiveMode::Recursive)?;
+//! # watcher.watch(Path::new("."), RecursiveMode::Recursive)?;
 //!
 //! use notify::Config;
 //! watcher.configure(Config::PreciseEvents(true))?;
@@ -77,6 +79,7 @@
 //!
 //! ```
 //! # use notify::{RecommendedWatcher, RecursiveMode, Result, Watcher};
+//! # use std::path::Path;
 //! #
 //! # fn main() -> Result<()> {
 //!       fn event_fn(res: Result<notify::Event>) {
@@ -86,10 +89,10 @@
 //!           }
 //!       }
 //!
-//!       let mut watcher1: RecommendedWatcher = Watcher::new_immediate(event_fn)?;
-//!       let mut watcher2: RecommendedWatcher = Watcher::new_immediate(event_fn)?;
-//! #     watcher1.watch(".", RecursiveMode::Recursive)?;
-//! #     watcher2.watch(".", RecursiveMode::Recursive)?;
+//!       let mut watcher1 = notify::recommended_watcher(event_fn)?;
+//!       let mut watcher2 = notify::recommended_watcher(event_fn)?;
+//! #     watcher1.watch(Path::new("."), RecursiveMode::Recursive)?;
+//! #     watcher2.watch(Path::new("."), RecursiveMode::Recursive)?;
 //! #
 //! #     Ok(())
 //! # }
@@ -100,7 +103,6 @@
 pub use config::{Config, RecursiveMode};
 pub use error::{Error, ErrorKind, Result};
 pub use event::{Event, EventKind};
-use std::convert::AsRef;
 use std::path::Path;
 
 #[cfg(all(target_os = "macos", feature = "macos_fsevent"))]
@@ -137,23 +139,16 @@ mod config;
 mod error;
 
 /// The set of requirements for watcher event handling functions.
-pub trait EventFn: 'static + Fn(Result<Event>) + Send {}
+pub trait EventFn: 'static + FnMut(Result<Event>) + Send {}
 
-impl<F> EventFn for F where F: 'static + Fn(Result<Event>) + Send {}
+impl<F> EventFn for F where F: 'static + FnMut(Result<Event>) + Send {}
 
 /// Type that can deliver file activity notifications
 ///
 /// Watcher is implemented per platform using the best implementation available on that platform.
 /// In addition to such event driven implementations, a polling implementation is also provided
 /// that should work on any platform.
-pub trait Watcher: Sized {
-    /// Create a new watcher in _immediate_ mode.
-    ///
-    /// Events will be sent using the provided `tx` immediately after they occur.
-    fn new_immediate<F>(event_fn: F) -> Result<Self>
-    where
-        F: EventFn;
-
+pub trait Watcher {
     /// Begin watching a new path.
     ///
     /// If the `path` is a directory, `recursive_mode` will be evaluated. If `recursive_mode` is
@@ -169,7 +164,7 @@ pub trait Watcher: Sized {
     ///
     /// [#165]: https://github.com/notify-rs/notify/issues/165
     /// [#166]: https://github.com/notify-rs/notify/issues/166
-    fn watch<P: AsRef<Path>>(&mut self, path: P, recursive_mode: RecursiveMode) -> Result<()>;
+    fn watch(&mut self, path: &Path, recursive_mode: RecursiveMode) -> Result<()>;
 
     /// Stop watching a path.
     ///
@@ -177,7 +172,7 @@ pub trait Watcher: Sized {
     ///
     /// Returns an error in the case that `path` has not been watched or if removing the watch
     /// fails.
-    fn unwatch<P: AsRef<Path>>(&mut self, path: P) -> Result<()>;
+    fn unwatch(&mut self, path: &Path) -> Result<()>;
 
     /// Configure the watcher at runtime.
     ///
@@ -218,9 +213,20 @@ pub type RecommendedWatcher = PollWatcher;
 /// _immediate_ mode.
 ///
 /// See [`Watcher::new_immediate`](trait.Watcher.html#tymethod.new_immediate).
-pub fn immediate_watcher<F>(event_fn: F) -> Result<RecommendedWatcher>
+pub fn recommended_watcher<F>(event_fn: F) -> Result<RecommendedWatcher>
 where
     F: EventFn,
 {
-    Watcher::new_immediate(event_fn)
+    // All recommended watchers currently implement `new`, so just call that.
+    RecommendedWatcher::new(event_fn)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_object_safe() {
+        let _watcher: &dyn Watcher = &NullWatcher;
+    }
 }