+++ CHANGELOG.md
@@ -19,7 +19,15 @@
 
 ## unreleased
 
-- FIX: Make StreamContextInfo Send to fix soundness issue [#325]
+- CHANGE: Make `Watcher` object safe [#336]
+- CHANGE: Change EventFn to take FnMut [#333]
+
+[#333]: https://github.com/notify-rs/notify/pull/333
+[#336]: https://github.com/notify-rs/notify/pull/336
+
+## 5.0.0-pre.10 (2021-06-04)
+
+- FIX: Make StreamContextInfo `Send` to fix soundness issue [#325]
 
 [#325]: https://github.com/notify-rs/notify/pull/325